feat: created validations for RevenueConfiguration creation,
